Move refactor
