+++ Descension/Assets/Scripts/Items/arrowsItem.cs
@@ -1,8 +1,8 @@
-using Assets.Scripts.Actor.Player;
-using Assets.Scripts.GUI.Controllers;
+using Actor.Player;
+using UI.Controllers;
 using UnityEngine;
 
-namespace Assets.Scripts.Items
+namespace Items
 {
     public class arrowsItem : MonoBehaviour
     {
