+++ Descension/Assets/Scripts/UI/Controllers/ButtonController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b3d8d7fa48e4cf43a5b67ae2c3364ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
