+++ Descension/Assets/Scripts/UI/Controllers/ButtonController/UINavButtonController.cs
@@ -1,6 +1,6 @@
-using Assets.Scripts.Util.Enums;
+using Util.Enums;
 
-namespace Assets.Scripts.GUI.Controllers.ButtonController
+namespace UI.Controllers.ButtonController
 {
     public class UINavButtonController : ButtonController
     {
