+++ Descension/Assets/Scripts/UI/MenuUI/UIController.cs
@@ -1,7 +1,7 @@
-using Assets.Scripts.Util.Enums;
+using Util.Enums;
 using UnityEngine;
 
-namespace Assets.Scripts.GUI.MenuUI
+namespace UI.MenuUI
 {
     public class UIController : MonoBehaviour
     {
