+++ Descension/Assets/Scripts/Util/AssetMenu/DeviceDisplayConfigurator.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
-using Assets.Scripts.Util.Structs;
+using Util.Structs;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
@@ -8,7 +8,7 @@ using UnityEngine.InputSystem;
  * Original code by Andy Touch (andyt@unity3d.com)
  * Source code https://github.com/UnityTechnologies/InputSystem_Warriors
  */
-namespace Assets.Scripts.Util.AssetMenu
+namespace Util.AssetMenu
 {
     [CreateAssetMenu(fileName = "Device Display Configurator",
         menuName = "Scriptable Objects/Device Display Configurator", order = 1)]
