+++ Descension/Assets/Scripts/Util/Constants/Direction.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Assets.Scripts.Util.Constants
+namespace Util.Constants
 {
     public static class Direction
     {
