+++ Descension/Assets/Scripts/Util/Structs/DeviceSet.cs
@@ -1,10 +1,10 @@
-﻿using Assets.Scripts.Util.AssetMenu;
+﻿using Util.AssetMenu;
 
 /*
  * Original code by Andy Touch (andyt@unity3d.com)
  * Source code https://github.com/UnityTechnologies/InputSystem_Warriors
  */
-namespace Assets.Scripts.Util.Structs
+namespace Util.Structs
 {
     [System.Serializable]
     public struct DeviceSet
